JBPM-6243 - Rename new controller module dropping the mgmt part
