+++ kie-server-parent/kie-server-controller/kie-server-controller-client/src/main/java/org/kie/server/controller/client/exception/KieServerControllerClientException.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2017 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.server.controller.client.exception;
+
+public class KieServerControllerClientException extends RuntimeException {
+
+    public KieServerControllerClientException() {
+    }
+
+    public KieServerControllerClientException(String message) {
+        super(message);
+    }
+
+    public KieServerControllerClientException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public KieServerControllerClientException(Throwable cause) {
+        super(cause);
+    }
+
+}
