--- kie-server-parent/kie-server-controller/kie-server-controller-websocket-common/src/main/java/org/kie/server/controller/websocket/common/WebSocketKieServerControllerClient.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2017 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.server.controller.websocket.common;
-
-import java.io.IOException;
-import java.util.List;
-import javax.websocket.Decoder;
-import javax.websocket.EncodeException;
-import javax.websocket.Encoder;
-
-import org.kie.server.controller.websocket.common.config.WebSocketClientConfiguration;
-import org.kie.server.controller.websocket.common.handlers.InternalMessageHandler;
-
-public interface WebSocketKieServerControllerClient {
-
-    void init(WebSocketClientConfiguration config);
-
-    void close();
-
-    void sendTextWithHandler(String content, InternalMessageHandler handler) throws IOException;
-
-    boolean isActive();
-}
