+++ kie-server-parent/kie-server-controller/pom.xml
@@ -23,7 +23,7 @@
     <module>kie-server-controller-websocket</module>
     <module>kie-server-controller-websocket-client</module>
     <module>kie-server-controller-websocket-common</module>
-    <module>kie-server-controller-mgmt-client</module>
+    <module>kie-server-controller-client</module>
   </modules>
 
 </project>
