+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-common/pom.xml
@@ -75,7 +75,7 @@
     </dependency>
     <dependency>
       <groupId>org.kie.server</groupId>
-      <artifactId>kie-server-controller-mgmt-client</artifactId>
+      <artifactId>kie-server-controller-client</artifactId>
     </dependency>
 
     <dependency>
