+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-common/src/main/java/org/kie/server/integrationtests/shared/KieServerSynchronization.java
@@ -49,7 +49,7 @@ import org.kie.server.client.RuleServicesClient;
 import org.kie.server.client.SolverServicesClient;
 import org.kie.server.client.UserTaskServicesClient;
 import org.kie.server.controller.api.model.spec.ServerTemplate;
-import org.kie.server.controller.management.client.KieServerMgmtControllerClient;
+import org.kie.server.controller.client.KieServerControllerClient;
 
 public class KieServerSynchronization {
 
@@ -244,7 +244,7 @@ public class KieServerSynchronization {
         });
     }
 
-    public static void waitForServerInstanceSynchronization(final KieServerMgmtControllerClient controllerClient, String serverTemplateId, final int numberOfExpectedServerInstances) throws Exception {
+    public static void waitForServerInstanceSynchronization(final KieServerControllerClient controllerClient, String serverTemplateId, final int numberOfExpectedServerInstances) throws Exception {
         waitForCondition(() -> {
             ServerTemplate serverTemplate = controllerClient.getServerTemplate(serverTemplateId);
             if (serverTemplate != null && serverTemplate.getServerInstanceKeys().size() == numberOfExpectedServerInstances) {
